Split out robots.txt, sitemap.xml, and status page
